delete useless stuff, make buttons to cycle through cards
